Bug 1081401: Remove unnecessary autoboxing (recalcitrant sync part). r=rnewman
UltraBlame original commit: 42758a959e470994cf11d4c68c422086740c4189

+++ mobile/android/base/sync/GlobalSession.java
@@ -1014,7 +1014,7 @@ public class GlobalSession implements HttpResponseObserver {
         if (version == null) {
           continue; 
         }
-        engineSettings = new EngineSettings(Utils.generateGuid(), version.intValue());
+        engineSettings = new EngineSettings(Utils.generateGuid(), version);
       } catch (NoSuchStageException e) {
